+++ pages/base_page.py
@@ -2,6 +2,8 @@ import time
 
 from selenium.webdriver.support.ui import Select
 
+from locators.locators_for_phone_page import *
+
 
 class BasePage:
 
@@ -51,3 +53,23 @@ class BasePage:
         self.driver_chrome.switch_to.window(tabs[1])
 
         return self.driver_chrome.current_url
+
+    def click_test(self, locator):
+        self.driver_chrome.execute_script("arguments[0].click();", self.find_element(locator))
+
+    def click_three_days(self):
+        """Кликаем по элементу"""
+
+        self.find_element(three_days_button).click()
+
+    def click_add_basket(self, locator):
+        """Кликаем по элементу"""
+
+        self.find_element(locator).click()
+
+    def click_clear_basket(self):
+        """Кликаем по элементу"""
+        self.find_element(clear_basket)
+
+    def click_original_phone(self):
+        self.find_element(button_original_phone).click()
